+++ src/Misc/Hooks.BugFixes.cpp
@@ -137,6 +137,9 @@ DEFINE_HOOK(0x702299, TechnoClass_ReceiveDamage_Debris, 0xA)
 
 	GET(TechnoClass* const, pThis, ESI);
 
+	if (pThis->Transporter)
+		return SkipGameCode;
+
 	const auto pType = pThis->GetTechnoType();
 
 	// Fix the debris count to be in range of Min, Max instead of Min, Max-1.
@@ -1371,7 +1374,19 @@ DEFINE_HOOK(0x6F4BB3, TechnoClass_ReceiveCommand_RequestUntether, 0x7)
 #pragma endregion
 
 // Fix the bug that techno in attack move will move to target if it cannot attack it
-DEFINE_JUMP(LJMP, 0x4D77BD, 0x4D769F)
+DEFINE_HOOK(0x4D77BD, FootClass_ObjectClickedAction_NoMove, 0x6)
+{
+	enum { Attack = 0x4D769F };
+
+	GET(ObjectClass*, pTarget, EBX);
+	const auto pTargetTechno = abstract_cast<TechnoClass*>(pTarget);
+
+	if (!pTargetTechno)
+		return 0;
+
+	GET(FootClass*, pThis, ESI);
+	return pThis->Owner->IsAlliedWith(pTargetTechno->Owner) ? 0 : Attack;
+}
 
 #pragma region JumpjetShadowPointFix